Explosion data too!
It's not much, but it's something. Also overhauled how explosion data is displayed. It should be much more accurate now.

Epicenters are now centered on the tile in which they originate, and the effect circle radius will extend to the furthest edge of the tile (if the explosion damage values were floored to begin with)